Move FakeWebSocketClient to its own class
